Set HIP_PATH based a search for hipcc
On Fedora hipcc is installed from an rpm to a system location /usr/bin/hipcc.
There is no /opt/rocm directory.

So if the user has not already set HIP_PATH, look for hipcc and if hipcc
is found either from the user's PATH or some other well known locations
set HIP_PATH based on the found location.

Signed-off-by: Tom Rix <trix@redhat.com>